+++ NEWS.md
@@ -1,8 +1,10 @@
 
-# unreleased
+
+# birdnetR 0.2.1
 
 - install the TensorFlow Metal Plugin for GPU support on Apple devices  #25
-- optionally use Apache Arrow to transfer data between R and Python to speed up inference with very low confidence thresholds #26
+- optionally use Apache Arrow to improve performance by reducing memory usage during data conversion and minimizing data
+  copying between R and Python. #26
 
 
 # birdnetR 0.2.0